Address first round of review comments
- default output backend/mode: zarr + distributed (config, driver, CLI)
- split writers.py into netcdf_writers.py and zarr_writers.py, keeping the
  shared surface (FieldWriter protocol, dimension names, attributes) in
  writers.py; writers are context managers now
- distributed.py: factor host (numpy) conversion of owner masks/global
  indices into _host_owner_data with the cupy/numpy rationale documented
- make total_wall_time a required argument of show_timer_report
- tests: use test_utils.assert_dallclose, match statements for backend
  dispatch, real types instead of object in the recording monitor,
  netcdf/gather pinned explicitly where tests assert netCDF artifacts
- ci: revert default.yml MODEL_MPI_SUBSETS to datatest
