+++ model/standalone_driver/tests/standalone_driver/unit_tests/test_driver_io.py
@@ -23,7 +23,8 @@ import pytest
 import xarray as xr
 
 from icon4py.model.common import dimension as dims, type_alias as ta
-from icon4py.model.common.grid import base, simple
+from icon4py.model.common.decomposition import definitions as decomposition_defs
+from icon4py.model.common.grid import base, simple, vertical as v_grid
 from icon4py.model.common.io import io as common_io
 from icon4py.model.common.states import (
     data as state_data,
@@ -219,13 +220,13 @@ def test_create_io_monitor_builds_single_field_group(
             self,
             *,
             config: common_io.IOConfig,
-            vertical_size: object,
-            horizontal_size: object,
+            vertical_size: v_grid.VerticalGrid,
+            horizontal_size: base.HorizontalGridSize,
             grid_file_name: pathlib.Path,
             grid_id: uuid.UUID,
             dtime: datetime.timedelta,
-            process_props: object,
-            decomposition_info: object,
+            process_props: decomposition_defs.ProcessProperties | None,
+            decomposition_info: decomposition_defs.DecompositionInfo | None,
         ) -> None:
             recorded["config"] = config
             recorded["grid_file_name"] = grid_file_name
@@ -247,9 +248,9 @@ def test_create_io_monitor_builds_single_field_group(
     field_group = config.field_groups[0]
     # default cadence: capture on every model step
     assert field_group.output_interval == 1
-    # default output setup: netCDF files, gathered to the root rank under MPI
-    assert field_group.backend == common_io.OutputBackend.NETCDF
-    assert field_group.mode == common_io.OutputMode.GATHER
+    # default output setup: zarr stores, every rank writing its own block under MPI
+    assert field_group.backend == common_io.OutputBackend.ZARR
+    assert field_group.mode == common_io.OutputMode.DISTRIBUTED
     # a single group holding all fields, prognostic + diagnostic, in one file
     assert list(field_group.variables) == driver_io.DEFAULT_OUTPUT_VARIABLES
     assert list(field_group.variables) == [
